+++ katana/payload.py
@@ -49,6 +49,7 @@ FIELD_MAPPINGS = {
     'config': 'c',
     'count': 'c',
     'cpu': 'c',
+    'caller': 'C',
     'calls': 'C',
     'complete': 'C',
     'command_reply': 'cr',
@@ -156,7 +157,21 @@ FIELD_MAPPINGS = {
     'timeout': 'x',
     'max_items': 'xi',
     'max_length': 'xl',
-}
+    }
+
+# Transport path for field that must be merged when service calls are made
+TRANSPORT_MERGEABLE_PATHS = (
+    'data',
+    'relations',
+    'links',
+    'calls',
+    'transactions',
+    'errors',
+    'body',
+    'files',
+    'meta/fallbacks',
+    'meta/properties',
+    )
 
 
 def get_path(payload, path, default=EMPTY, mappings=None, delimiter=SEP):