+++ CHANGELOG.md
@@ -4,5 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.1] - 2017-03-07
+### Added
+- Transport merge support for run-time calls.
+
+### Changed
+- Updated payload argument names for "action" and "callee" to be consistent
+  between transactions and calls in API `Action` class.
+
 ## [1.0.0] - 2017-03-01
 - Initial release